# skillhub_back/storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillHub API",
        "description": "API REST SkillHub - Formations et JWT",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "paths": {
        "/auth/inscription": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscription",
                "description": "Inscription d'un nouvel utilisateur.",
                "operationId": "f2d03426fa2606c0b7bfcdb4b9e9e578",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mot_de_passe",
                                    "nom",
                                    "role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "participant",
                                            "formateur"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/auth/connexion": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connexion",
                "description": "Connexion.",
                "operationId": "18af88b9c23de450cabcec8616eee63f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mot_de_passe"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "mot_de_passe": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie - retourne token JWT"
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/auth/deconnexion": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Déconnexion",
                "description": "Déconnexion (invalide le token).",
                "operationId": "acfbba3f95e35e0406e164ac0b6f44c8",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Utilisateur connecté",
                "description": "Utilisateur connecté.",
                "operationId": "7edae2cdbaafc10dff8c9c4198b320a2",
                "responses": {
                    "200": {
                        "description": "Données utilisateur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rafraîchir le token",
                "description": "Rafraîchir le token.",
                "operationId": "e7ab7446393758bb71fd1946a349e500",
                "responses": {
                    "200": {
                        "description": "Nouveau token"
                    },
                    "401": {
                        "description": "Token invalide ou expiré"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}